remove MSVC compiler warning (required a type cast)
svn path=/trunk/; revision=11741

+++ epan/dissectors/packet-dhcp-failover.c
@@ -835,7 +835,7 @@ dissect_dhcpfo(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree)
 
         			case DHCP_FO_PD_MESSAGE_DIGEST:
 
-					message_digest_type = tvb_get_ntohs(tvb,helpliste->actualpoffset+4);
+					message_digest_type = (guint8) tvb_get_ntohs(tvb,helpliste->actualpoffset+4);
 					if(message_digest_type == 1)
 					{
 						proto_item_append_text(oi, ", HMAC-MD5");
